tracker/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
apscheduler==3.10.4
requests==2.32.3
gitpython==3.1.43
